requirements.txt
openai>=1.14.0
azure-cognitiveservices-vision-computervision>=0.9.0
azure-ai-textanalytics>=5.3.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pytest>=8.0.0
pytest-mock>=3.12.0
